xagent/xagent/src/services/document/processors/python/requirements.txt
python-magic==0.4.27
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
openpyxl==3.1.2
pandas==2.1.3
pillow==10.1.0
numpy==1.26.2
pdfminer.six==20221105